requirements.txt
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
python-docx>=1.1
